design_patterns/structural/facade_pattern/facade_pattern.py
# In summary, the Car class acts as a facade, hiding the complexity of the subsystems
# and providing a simple interface to the client, which in this case is the FacadePattern class.
# This makes the client code easier to read and maintain.

class IntakeSystem:
    def open_throttle(self):
        pass


class FuelSystem:
    def inject_fuel(self):
        pass


class IgnitionSystem:
    def create_spark(self):
        pass


class ElectricalSystem:
    def switch_on(self):
        pass


class Car:
    """
    The Car class acts as a facade, hiding the complexity of the subsystems
    and providing a simple interface to the client, which in this case is the FacadePattern class.
    """
    def __init__(self):
        # Subsystems
        self.intake_system = IntakeSystem()
        self.fuel_system = FuelSystem()
        self.ignition_system = IgnitionSystem()
        self.electrical_system = ElectricalSystem()

    def start_engine(self):
        # simplified method to start the engine as a facade
        self.electrical_system.switch_on()
        self.intake_system.open_throttle()
        self.fuel_system.inject_fuel()
        self.ignition_system.create_spark()


if __name__ == "__main__":
    car = Car()
    car.start_engine()
